ESMILES grammar rewritten
The grammar was completely rewritten as the original did not provided
enough means to add bonds to branches (i.e. specify through which kind
of bond would the branch be attached to the main chain).

Rules for bonded branches were added.

Simple example of acetone was added do demonstrate that double bond is
recognised for a branch.